resolve vc++ warnings

+++ lib/clib/hash.c
@@ -16,8 +16,8 @@
 unsigned int hash(unsigned char *data, size_t len)
 {
     const char *str = (const char *)data;
-    unsigned int hash = len;
-    unsigned int i = 0;
+    unsigned hash = (unsigned)len;
+    unsigned i = 0;
     for (i = 0; i < len; ++str, ++i) {
         hash = ((hash << 5) ^ (hash >> 27)) ^ (*str);
     }
